// src/core/enums/contract-type.enum.ts
export enum ContractScheduleType {
  FIXED = "fixed",
  RANGE = "range",
  MAXIMUM = "maximum",
  MINIMUM = "minimum",
}

export enum ContractDurationUnit {
  DAYS = "days",
  WEEKS = "weeks",
  MONTHS = "months",
}

export enum ContractType {
  FULL_TIME = "Temps plein",
  PART_TIME = "Temps partiel",
  CDI = "CDI",
  INTERIM = "Interim",
  CDD = "CDD",
  FREELANCE = "Freelance",
  INTERNSHIP = "Stage",
  APPRENTICESHIP = "Apprentissage",
}